+++ lib/tescon/analyzer.rb
@@ -8,6 +8,7 @@ require_relative "rules/expect_eq"
 require_relative "rules/expect_include"
 require_relative "rules/expect_match"
 require_relative "rules/expect_be_present"
+require_relative "rules/expect_be_valid"
 require_relative "rules/expect_raise_error"
 require_relative "rules/factory_bot"
 require_relative "rules/is_expected_eq"
@@ -41,6 +42,7 @@ module Tescon
       Rules::ExpectBeNil.new,
       Rules::ExpectBeTruthy.new,
       Rules::ExpectBePresent.new,
+      Rules::ExpectBeValid.new,
       Rules::ExpectEq.new,
       Rules::ExpectInclude.new,
       Rules::ExpectMatch.new,